% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jing.R
\name{jing}
\alias{jing}
\title{Jing - and our basic communications}
\usage{
jing()
}
\value{
A sentence from us.
}
\description{
Jing, if you ever miss us - just run yourself in the code. Someone will be there.
}
\examples{
#Oh come on - just try it.
jing()
}

