% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jing.R
\name{jing_leaves}
\alias{jing_leaves}
\title{This function needs to be run whenever you leave the office.}
\usage{
jing_leaves(time = 6)
}
\arguments{
\item{Whenever}{you leave, Jing!}
}
\value{
Well, it returns our basic conversation every evening
}
\description{
Really, this is true. Dont go before 6pm! I will be still in my office, and you know it!
}
\examples{
jing_leaves(time = 5)
}

