% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jing.R
\name{jing_support}
\alias{jing_support}
\title{Everyone needs a little support sometimes}
\usage{
jing_support()
}
\value{
A little support for Jing, from Germany!
}
\description{
Everyone needs a little support sometimes
}
\examples{
# Just do it
jings_support()
}

